fix: store correct gridSnapshot and fix win_level column type
- SpinUseCase: gridSnapshot was storing outcome.grid (undefined); now stores
  outcome.baseSpins[0].grid which is the actual 2D symbol grid
- Migration: change win_level from INTEGER to NUMERIC(18,4) — PAYTABLE_SCALE=3.622
  produces non-integer win multipliers causing DB insert failures

+++ supabase/migrations/20260328000001_initial_schema.sql
@@ -154,7 +154,7 @@ CREATE TABLE spin_logs (
     -- Currency & normalization (§9-M)
     currency        CHAR(3)         NOT NULL DEFAULT 'USD',
     bet_level       INTEGER         NOT NULL CHECK (bet_level > 0),
-    win_level       INTEGER         NOT NULL DEFAULT 0,
+    win_level       NUMERIC(18,4)   NOT NULL DEFAULT 0,
     base_unit       NUMERIC(18,8)   NOT NULL,
     player_bet      NUMERIC(18,4)   NOT NULL,
     player_win      NUMERIC(18,4)   NOT NULL DEFAULT 0,
